requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
python-dotenv
joblib
langgraph
langchain
groq
openai
chromadb
